refactor: useOrder, useCustomerCharge 수정
1. ContextProvider 내에서 사용하던 useorder, useCustomerCharge 제거
2. useContext를 사용하는 useOrder, useCustomerCharge를 새로 작성
