+++ src/components/DrinkSection/DrinkPickupSection.tsx
@@ -1,6 +1,7 @@
-import React, { useContext } from "react";
-import { OrderContext } from "../../context/OrderContext";
+import React from "react";
+
 import { ORDER_PROGRESS } from "../../constants";
+import { useOrder } from "./../../context/OrderContext";
 
 interface Props {
   ingredientList: string[];
@@ -8,7 +9,7 @@ interface Props {
 }
 
 const DrinkPickupSection = ({ ingredientList, clearIngredientList }: Props) => {
-  const { order, updateOrderStateToPending } = useContext(OrderContext);
+  const { order, updateOrderStateToPending } = useOrder();
 
   const handleClickDrinkPickupButton = () => {
     updateOrderStateToPending();
